+++ changelog.md
@@ -4,6 +4,7 @@
     * E.g. Vindication no longer shows accumulated damage
 * Veining and Tunneling Runes' mining speed penalty is now configurable
 * Slightly increased Thorns chance to trigger
+* Increased Atmospheric under rain durability consumption reduction
 * Fixed Feather Falling applicable to any armor piece
 
 # 1.0.1.3-beta